+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: gchalifo <gchalifo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 08:22:26 by gchalifo          #+#    #+#             */
-/*   Updated: 2023/02/28 09:50:16 by gchalifo         ###   ########.fr       */
+/*   Updated: 2023/03/02 11:02:03 by gchalifo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	char	*s2;
 	size_t	i;
 
-	if (len > ft_strlen(s))
+	if (len > (size_t) ft_strlen(s))
 		len = ft_strlen(s);
 	s2 = (char *) malloc((len + 1) * sizeof(char));
 	if (s2 == NULL)
 		return (NULL);
 	i = 0;
-	while (i < len && (start + i < ft_strlen(s)))
+	while (i < len && (start + (unsigned int)i < (unsigned int) ft_strlen(s)))
 	{
 		s2[i] = s[i + start];
 		i++;
